Complete FemElementMesh loading pipeline Update unit tests for new loading method Update PlyReaders to use unsigned int instead of size_t

+++ SurgSim/Physics/FemElement1DMeshPlyReaderDelegate.h
@@ -155,7 +155,7 @@ private:
 	bool m_hasBoundaryConditions;
 
 	/// Internal data to receive the "boundary_condition" element
-	size_t m_boundaryConditionData;
+	unsigned int m_boundaryConditionData;
 
 	std::shared_ptr<FemElement1DMesh> m_mesh;
 };
